[EXTRA] Refactor and cleanup the map array container
This PR runs the following refactor of the map and array containers.

- Remove InplaceArrayBase since it is not as useful
- Extract MapObj into map_base.h so it can be used for other different subclass and variant of map if needed.
- Add extra helper methods such as clean

We also reorganized SmallBaseMapObj after DenseMapObj as SmallBaseMapObj is a specialization, which
can help possible future refactors to make SmallBaseMapObj be aware of DenseMapObj.
